Integrate search into HomeView and remove SearchView
