--- MEAL AI/Sources/Views/Search/SearchView.swift
@@ -1,106 +1,0 @@
-
-import SwiftUI
-import UIKit
-
-struct SearchView: View {
-    @State private var query: String = ""
-    @State private var showingScanner = false
-    @State private var showingCamera = false
-    @State private var isLoading = false
-    @State private var errorMessage: String?
-    @State private var result: StageMealResult?
-    @State private var currentImage: UIImage?
-    @State private var currentImages: [UIImage] = []
-
-    @ObservedObject private var favs = FavoritesStore.shared
-
-    var body: some View {
-        ZStack {
-        VStack(spacing: 12) {
-            HStack {
-                TextField("Hae ruokaa nimellä…", text: $query, onCommit: runTextSearch)
-                    .textFieldStyle(.roundedBorder)
-            }.padding(.horizontal)
-
-            HStack {
-                Button { showingScanner = true } label: {
-                    Label("Viivakoodi", systemImage: "barcode.viewfinder")
-                }.buttonStyle(.bordered)
-                Button {
-                    showingCamera = true
-                    currentImage = nil
-                    currentImages = []
-                } label: {
-                    Label("Kamera", systemImage: "camera")
-                }.buttonStyle(.bordered)
-                Spacer()
-            }
-            .padding(.horizontal)
-
-            if let e = errorMessage { Text(e).foregroundColor(.red).padding(.horizontal) }
-
-            if !favs.items.isEmpty {
-                List {
-                    Section("Suosikit") {
-                        ForEach(favs.items) { f in
-                            Button { result = f.result } label: {
-                                HStack {
-                                    Text(f.name)
-                                    Spacer()
-                                    let t = f.result.analysis
-                                    Text("\(Int(t.carbs_g))g C").foregroundStyle(.secondary)
-                                }
-                            }
-                        }
-                    }
-                }
-            } else {
-                Spacer()
-            }
-        }
-        .navigationTitle("Ruokahaku")
-        .sheet(isPresented: $showingScanner) {
-            BarcodeScanView { code in
-                showingScanner = false
-                currentImage = nil
-                Task { await run(.barcode(code)) }
-            }
-        }
-        .sheet(isPresented: $showingCamera) {
-            CameraView { datas in
-                showingCamera = false
-                currentImages = datas.compactMap { UIImage(data: $0) }
-                currentImage = currentImages.first
-                Task { await run(.images(datas)) }
-            }
-        }
-        .sheet(item: $result) { r in ResultView(result: r, image: currentImage) }
-
-        if isLoading {
-            AnalysisOverlayView(onCancel: { isLoading = false })
-        }
-        }
-    }
-
-    private func runTextSearch() {
-        currentImage = nil
-        Task { await run(.text(query)) }
-    }
-
-    private func run(_ input: FoodSearchRouter.Input) async {
-        guard UserDefaults.standard.foodSearchEnabled else {
-            self.errorMessage = "Food Search ei ole päällä (Asetukset → Food Search)."
-            return
-        }
-        errorMessage = nil
-        isLoading = true
-        defer { isLoading = false }
-        do {
-            let stage = try await FoodSearchRouter.shared.run(input)
-            self.result = stage
-        } catch {
-            self.errorMessage = error.localizedDescription
-        }
-    }
-
-}
